split up hpp tests
