+++ src/nanoarrow/hpp/array_stream_test.cc
@@ -15,14 +15,50 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include <array>
-#include <cerrno>
-
 #include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
 #include "nanoarrow/nanoarrow.hpp"
-#include "nanoarrow/nanoarrow_gtest_util.hpp"
-#include "nanoarrow/nanoarrow_testing.hpp"
 
-TEST(HppArrayStream, Test) { EXPECT_EQ(1, 1); }
+using testing::ElementsAre;
+
+TEST(HppArrayStream, EmptyArrayStream) {
+  nanoarrow::UniqueSchema schema;
+  struct ArrowArray array;
+
+  nanoarrow::UniqueSchema schema_in;
+  EXPECT_EQ(ArrowSchemaInitFromType(schema_in.get(), NANOARROW_TYPE_INT32), NANOARROW_OK);
+
+  nanoarrow::UniqueArrayStream array_stream;
+  nanoarrow::EmptyArrayStream(schema_in.get()).ToArrayStream(array_stream.get());
+
+  EXPECT_EQ(ArrowArrayStreamGetSchema(array_stream.get(), schema.get(), nullptr),
+            NANOARROW_OK);
+  EXPECT_STREQ(schema->format, "i");
+  EXPECT_EQ(ArrowArrayStreamGetNext(array_stream.get(), &array, nullptr), NANOARROW_OK);
+  EXPECT_EQ(array.release, nullptr);
+  EXPECT_STREQ(ArrowArrayStreamGetLastError(array_stream.get()), "");
+}
+
+TEST(HppArrayStream, VectorArrayStream) {
+  nanoarrow::UniqueArray array_in;
+  EXPECT_EQ(ArrowArrayInitFromType(array_in.get(), NANOARROW_TYPE_INT32), NANOARROW_OK);
+  EXPECT_EQ(ArrowArrayStartAppending(array_in.get()), NANOARROW_OK);
+  EXPECT_EQ(ArrowArrayAppendInt(array_in.get(), 1234), NANOARROW_OK);
+  EXPECT_EQ(ArrowArrayFinishBuildingDefault(array_in.get(), nullptr), NANOARROW_OK);
+
+  nanoarrow::UniqueSchema schema_in;
+  EXPECT_EQ(ArrowSchemaInitFromType(schema_in.get(), NANOARROW_TYPE_INT32), NANOARROW_OK);
+
+  nanoarrow::UniqueArrayStream array_stream;
+  nanoarrow::VectorArrayStream(schema_in.get(), array_in.get())
+      .ToArrayStream(array_stream.get());
+
+  nanoarrow::ViewArrayStream array_stream_view(array_stream.get());
+  for (ArrowArray& array : array_stream_view) {
+    EXPECT_THAT(nanoarrow::ViewArrayAs<int32_t>(&array), ElementsAre(1234));
+  }
+  EXPECT_EQ(array_stream_view.count(), 1);
+  EXPECT_EQ(array_stream_view.code(), NANOARROW_OK);
+  EXPECT_STREQ(array_stream_view.error()->message, "");
+}
